Renamed polarispp files/library to polaris_client.
These names do not conflict with the C library.
